Added data dictionary edit function

+++ dd-delete.php
@@ -1,9 +1,10 @@
 <?php
 include 'config.php';
+if (isset($_GET['delete-id'])) {
     $delete_id = $_GET['delete-id'];
     $table = $_GET['tbname'];
     $delete_sql = "DELETE FROM $table WHERE uid = $delete_id";
     pg_query($delete_sql);
-
+}
 
 ?>
